// renderer/types/itemType.d.ts
export type itemType = {
  barcode: string;
  name: string;
  qty: number;
  sellPrice1: number;
  sellPrice2?: number;
  sellPrice3?: number;
  sellPrice4?: number;
  sellPrice5?: number;
  sellCondition2?: number;
  sellCondition3?: number;
  sellCondition4?: number;
  sellCondition5?: number;
  getPrice: number;
  isActive: boolean;
  unit: string;
};